Fix saveNewMod
